+++ NN_2L2H2O_poly2d_CPU/src/fullTester.cpp
@@ -12,11 +12,11 @@
 #include <math.h>
 #include <iterator>
 
-#include "readGparams.h"
-#include "atomTypeID.h"
+#include "readGparams_v2.h"
+#include "atomTypeID_v2.h"
 #include "utility.h"
 #include "timestamps.h"
-#include "Gfunction.h"
+#include "Gfunction_v2.h"
 #include "network.h"
 
 #define INFILE1     "/server-home1/ndanande/Documents/mbpol_cuda_devel/NN_2L2H2O_poly2d_CPU/34.hdf5"     // HDF5 files for Layer Data
@@ -47,7 +47,6 @@ const double THREDHOLD_MAX_VALUE = 60.0;
 int main(int argc, char** argv){ 
 
      cout << "Usage:  THIS.EXE  XYZ_FILE " 
-          << "[-" << FLAG_COLUMN_INDEX_FILE  << "=NONE]"  
           << "[-" << FLAG_PARAM_FILE         << "=H_rad|H_ang|O_rad|O_ang]"
           << "[-" << FLAG_ATOM_ORDER_FILE    << "=NONE]"
 		<< "[-" << FLAG_GFN_OUTPUT_ENABLE 	<< "=0]"		
@@ -59,9 +58,6 @@ int main(int argc, char** argv){
 
      Gfunction_t<double> gf;     // the G-function
    
-     // column index file
-     string colidxfile;
-     getCmdLineArgumentString(argc, (const char **)argv, FLAG_COLUMN_INDEX_FILE, colidxfile);
           
      // parameter file
      string paramfile;
@@ -73,44 +69,50 @@ int main(int argc, char** argv){
 
 	//Gfn output enable
 	int gfnOut = getCmdLineArgumentInt(argc, (const char **)argv, FLAG_GFN_OUTPUT_ENABLE);     
-       
+
 	//make G function with xyz file
-     gf.make_G_XYZ(argv[1],colidxfile.c_str(), paramfile.c_str(), ordfile.c_str());
+     gf.make_G_XYZ(argv[1], paramfile.c_str(), ordfile.c_str());
 
      // normalize G by the maximum value of first 90% dimers
      //gf.norm_G_by_maxabs_in_first_percent(TRAIN_PERCENT);    
      // results saved in gf.G which is a map<string:atom_idx, double**>
-     
+
 	//define number of samples
-	size_t N = gf.ndimers;
-	int testIndexStart = N*TRAIN_PERCENT;
-	N -= testIndexStart;		//effective new N(for test size) after 90% used for training
-	
+	size_t N = gf.NCluster;
+	//int testIndexStart = N*TRAIN_PERCENT;
+	//N -= testIndexStart;		//effective new N(for test size) after 90% used for training
+
+	cout<<"Something"<<endl;
 	//define number of atoms
 	size_t numAtoms = gf.G.size();
 
+
 	//init memory to store test set gfn Outputs
 	double ** X = new double * [numAtoms];
-	
+
 	//store inputDimension for each seperate atom
 	size_t * inputDim = new size_t[numAtoms];
 
 		
+	
 	int i=0;	//iterating variable
+
+	cout<<"Begin, numAtoms: " << numAtoms<<" Num Clusters: " << N<<endl;
 	//put G function outputs into 2d array for usage in NN(only the test portion)
  	for(auto it=gf.G.begin(); it!=gf.G.end(); it++){
-		//string atom = gf.model.atoms[it->first]->name;
-		string atom_type = gf.model.atoms[it->first]->type;
-		inputDim[i] = gf.G_param_max_size[atom_type];
+		inputDim[i] = gf.G_param_max_size[gf.model.TYPE_EACHATOM[i]];
+		cout<<"Input Dim " << i << " : " << inputDim[i]<<endl; 
 		X[i] = new double[inputDim[i]*N];
 		for(int ii=0; ii<N;ii++){
 			for(int jj=0;jj<inputDim[i];jj++){
-				X[i][ii*inputDim[i] + jj] = it->second[jj][ii+testIndexStart];
+				X[i][ii*inputDim[i] + jj] = (*it)[ii][jj];
 			}
 		}
 		i++;
 	}
 	
+
+
 	//if user specified to output intermediate Gfn outputs to files, do so
 	if(gfnOut != 0){
 		i=0;		//reset iterating variable
@@ -118,15 +120,14 @@ int main(int argc, char** argv){
 		string gfuncOutPrefix = "gfn_o_";
 		string gfuncOutPath = "";
 		for(auto it=gf.G.begin();it!=gf.G.end();it++){
-			string atom = gf.model.atoms[it->first]->name;
-			string atom_type = gf.model.atoms[it->first]->type;
-			atom.erase(remove(atom.begin(), atom.end(), ')'),atom.end());	//get rid of () for better file name format
-			atom.erase(remove(atom.begin(), atom.end(), '('),atom.end());
-			gfuncOutPath = gfuncOutPrefix+atom;
+			string atomType = std::to_string(gf.model.TYPE_EACHATOM[i]);
+			string atomNum = std::to_string(i);
+			gfuncOutPath = gfuncOutPrefix+atomType+"_"+atomNum;
 			gfuncOut.open(gfuncOutPath);
-			for(int j=0;j<gf.ndimers;j++){
+			cout<<"inputDim " << atomType<<atomNum<<" inputDim[i] " << "N: "<< N<<endl;
+			for(int j=0;j<N;j++){
 				for(int k=0;k<inputDim[i];k++){
-					gfuncOut<<setprecision(18)<<scientific<<it->second[k][j]<<" ";
+					gfuncOut<<setprecision(18)<<scientific<<(*it)[j][k]<<" ";
 				}
 				gfuncOut<<endl;
 			}
@@ -135,6 +136,7 @@ int main(int argc, char** argv){
 		}
 	}
 
+     cout<<("TEST 5")<<endl; 
 	cout<<"Running Network"<<endl;
 	runtester<double>(INFILE1,CHECKCHAR2,X,numAtoms,N,inputDim);
