+++ NN_2L2H2O_poly2d_CPU/src/utility.cpp
@@ -5,7 +5,6 @@
 #include <string.h>
 #include <stdlib.h>
 #include <stdio.h>
-#include <math.h>
 
 #include <iostream>
 #include <sstream>
@@ -16,6 +15,7 @@
 #include <string>
 
 #include "utility.h"
+//#include "atomTypeID.h"
 
 // Define the cblas library 
 #ifdef _USE_GSL
@@ -39,6 +39,22 @@
 using namespace std;
 
 
+
+
+template<>
+double return_a_number<double>(string _string){
+     return  stod ( _string );     
+}
+
+template<>
+float return_a_number<float>(string _string){
+     return  stof ( _string );     
+}
+
+template<>
+int return_a_number<int>(string _string){
+     return  stoi  ( _string );     
+}
 
 //==============================================================================
 //
@@ -117,6 +133,7 @@ size_t get_count_by_percent(size_t src_count, double percentage){
 
 template<>
 void get_max_each_row<double>(double*& rst, double* src, size_t src_rows, size_t src_cols, long int col_start, long int col_end){
+     if(col_start < 0) col_start = src_cols + col_start;
      if(col_end < 0) col_end = src_cols + col_end ;  // change negative column index to positive
      if(rst == nullptr) rst = new double[src_rows]();   
       
@@ -132,6 +149,7 @@ void get_max_each_row<double>(double*& rst, double* src, size_t src_rows, size_t
 
 template<>
 void get_max_each_row<float>(float*& rst, float* src, size_t src_rows, size_t src_cols, long int col_start, long int col_end){
+     if(col_start < 0) col_start = src_cols + col_start;
      if(col_end < 0) col_end = src_cols + col_end ;  // change negative column index to positive
      if(rst == nullptr) rst = new float[src_rows]();          
      
@@ -147,6 +165,7 @@ void get_max_each_row<float>(float*& rst, float* src, size_t src_rows, size_t sr
 
 template<>
 void norm_rows_in_mtx_by_col_vector(double* src_mtx, size_t src_rows, size_t src_cols, double* scale_vec, long int col_start, long int col_end){
+     if(col_start < 0) col_start = src_cols + col_start;
      if(col_end < 0) col_end = src_cols + col_end ;  // change negative column index to positive
      // scale each row (from col_start[0,1,2...] to col_end[ ...-3, -2,-1]) in a matrix by a column vector
      #ifdef _OPENMP
@@ -159,6 +178,7 @@ void norm_rows_in_mtx_by_col_vector(double* src_mtx, size_t src_rows, size_t src
 
 template<>
 void norm_rows_in_mtx_by_col_vector(float* src_mtx, size_t src_rows, size_t src_cols, float* scale_vec, long int col_start, long int col_end){
+     if(col_start < 0) col_start = src_cols + col_start;
      if(col_end < 0) col_end = src_cols + col_end ;  // change negative column index to positive
      // scale each row (from col_start[0,1,2...] to col_end[ ...-3, -2,-1]) in a matrix by a column vector
      #ifdef _OPENMP
@@ -296,317 +316,3 @@ bool getCmdLineArgumentString(const int argc, const char **argv,
 
 
 
-
-
-
-//templated function definitions//
-
-//==============================================================================
-//
-// Initialize a matrix in consecutive memory
-template <typename T>
-bool init_mtx_in_mem(T** & data, size_t rows, size_t cols){
-     try{          
-          //clearMemo<T>(data);
-          if( rows*cols >0) {          
-               T * p = new T[rows*cols]();
-               data = new T* [rows];
-               #ifdef _OPENMP
-               #pragma omp parallel for shared(data, p, rows, cols)
-               #endif 
-               for(int ii=0; ii<rows; ii++){                    
-                    data[ii]=p+ii*cols;     
-                    memset(data[ii], 0, sizeof(T)*cols);       
-               }          
-          }     
-          return true;
-     } catch (...){
-          clearMemo<T>(data);   
-          return false;
-     }
-}
-
-
-//==============================================================================
-//
-// Check if a string is a float number
-template <typename T>
-bool IsFloat( std::string myString ) {
-    std::istringstream iss(myString);
-    T f;
-    iss >> std::skipws >> f; // skipws ignores leading whitespace
-    // Check the entire string was consumed and if either failbit or badbit is set
-    return iss.eof() && !iss.fail(); 
-}
-
-
-//==============================================================================
-//
-// Memory management functions
-//
-template <typename T>
-void clearMemo(T** & data){
-     if (data != nullptr) {
-          if (data[0] != nullptr)  delete[] data[0];   
-          delete[] data;
-     };
-     data = nullptr;
-     return;
-}
-
-template <typename T>
-void clearMemo(std::vector<T**> & data){
-     for (auto it=data.rbegin(); it!=data.rend(); it++){
-          clearMemo<T>(*it);
-          data.pop_back();
-     }
-     return;
-}
-
-
-template <typename T>
-void clearMemo(std::map<std::string, T**> & data){
-     for (auto it=data.begin(); it!=data.end(); it++){
-          clearMemo<T>(it->second);
-          it->second = nullptr;
-     }
-     return;
-}
-
-
-//==============================================================================
-//
-// Read in a 2D array from file and save to  **data / rows / cols
-template <typename T>
-int read2DArrayfile(T** & data, size_t& rows, size_t& cols, const char* file, int titleline=0){
-    try { 
-          
-          clearMemo<T>(data);
-          
-          std::ifstream ifs(file);
-          std::string line;
-          matrix_by_vector_t<T> mtx;
-          
-          for (int i=0; i < titleline; i++){          
-               getline(ifs,line);
-          }
-          std::vector<T> onelinedata;
-          while(getline(ifs, line)){          
-               char* p = &line[0u];
-               char* end;              
-               onelinedata.clear();                              
-               for( T d = strtod(p, &end); p != end; d = strtod(p, &end) ) {
-                    p = end;                    
-                    onelinedata.push_back(d);           
-               };               
-               if(onelinedata.size()>0) mtx.push_back(onelinedata);              
-          }          
-
-          rows=mtx.size();
-          if (rows > 0){
-               cols=mtx[0].size();
-               
-               init_mtx_in_mem<T>(data, rows, cols);  
-               
-               #ifdef _OPENMP
-               #pragma omp parallel for simd shared(data, mtx, rows)
-               #endif                                      
-               for(int ii=0; ii<rows; ii++){
-                    copy(mtx[ii].begin(), mtx[ii].end(), data[ii]);       
-               }          
-          } else {
-               std::cout << " No Data is read from file as 2D array" << std::endl;
-          }
-          mtx.clear();
-          return 0;                    
-    } catch (const std::exception& e) {
-        std::cerr << " ** Error ** : " << e.what() << std::endl;
-        return 1;
-    }
-};
-
-template <typename T>
-int read2DArray_with_max_thredhold(T** & data, size_t& rows, size_t& cols, const char* file, int titleline, 
-                    int thredhold_col, T thredhold_max){
-    try { 
-          
-          clearMemo<T>(data);
-          
-          std::ifstream ifs(file);
-          std::string line;
-          matrix_by_vector_t<T> mtx;
-          
-          for (int i=0; i < titleline; i++){          
-               getline(ifs,line);
-          }
-          std::vector<T> onelinedata;
-          while(getline(ifs, line)){          
-               char* p = &line[0u];
-               char* end;              
-               onelinedata.clear();                              
-			
-               for( T d = strtod(p, &end); p != end; d = strtod(p, &end) ) {
-                    p = end;                  
-                    onelinedata.push_back(d);           
-               };               
-               if (onelinedata.size()>0) {   
-                    int checkcol = onelinedata.size() ;                                                        
-                    if ( thredhold_col >=0) {           
-                         // when thredhold_index is non-negative, check the colum VS max
-                         checkcol = thredhold_col;
-                    } else {
-                         // when thredhold_index is negative, check the column from the end VS max
-                         checkcol += thredhold_col;                                        
-                    }                         
-                    if (onelinedata[checkcol] > thredhold_max) {
-                         continue;  // If the data exceeds thredhold, ignore this line.
-                    }
-                    mtx.push_back(onelinedata);                               
-               }                           
-          }          
-
-          rows=mtx.size();
-          if (rows > 0){
-               cols=mtx[0].size();
-               
-               init_mtx_in_mem<T>(data, rows, cols);  
-               
-               #ifdef _OPENMP
-               #pragma omp parallel for simd shared(data, mtx, rows)
-               #endif                                      
-               for(int ii=0; ii<rows; ii++){
-                    copy(mtx[ii].begin(), mtx[ii].end(), data[ii]);       
-               }          
-          } else {
-               std::cout << " No Data is read from file as 2D array" << std::endl;
-          }
-          mtx.clear();
-          return 0;                    
-    } catch (const std::exception& e) {
-        std::cerr << " ** Error ** : " << e.what() << std::endl;
-        return 1;
-    }
-};
-
-//========================================================================================
-// 2D array transpose
-//
-template <typename T>
-void transpose_mtx(T** & datdst,  T** datrsc,  size_t nrow_rsc, size_t ncol_rsc)
-{
-     try{ 
-     
-          if ( datdst== nullptr) init_mtx_in_mem<T>(datdst, ncol_rsc, nrow_rsc);
-          
-          // best way to transpose a c++ matrix is copying by row    
-          #ifdef _OPENMP
-          #pragma omp parallel for simd shared(datrsc, datdst, nrow_rsc, ncol_rsc)
-          #endif      
-          for(int irow = 0; irow< nrow_rsc; irow++){          
-               for(int icol=0; icol < ncol_rsc; icol++){
-                    datdst[icol][irow] = datrsc[irow][icol];      // not a smart way.
-               }
-          }          
-          
-     } catch (const std::exception& e) {
-        std::cerr << " ** Error ** : " << e.what() << std::endl;
-     }
-};
-
-
-//===============================================================================                                     
-//
-// Matrix normalization utility functions
-template<typename T>
-void get_max_each_row(T*& rst,  T* src,  size_t src_rows, size_t src_cols, long int col_start, long int col_end){ 
-          if(col_end < 0) col_end = src_cols + col_end ;  // change negative column index to positive        
-          if(rst == nullptr) rst = new T[src_rows]();               
-          #ifdef _OPENMP
-          #pragma omp parallel for simd shared(src, rst, src_rows, src_cols, col_start, col_end)
-          #endif   
-		int idxlast = 0;
-          for(size_t ii=0 ; ii< src_rows ; ii++ ){
-               for(size_t jj=col_start; jj<= col_end ; jj++){
-                    if ( rst[ii] < fabs(src[ii*src_cols + jj]) ) {
-                         rst[ii] = fabs(src[ii*src_cols + jj]);
-					if(ii==(src_rows-1)){
-						idxlast = jj;
-					}
-                    };
-               }
-          };  
-};
-
-
-template<typename T>
-void norm_rows_in_mtx_by_col_vector(T* src_mtx, size_t src_rows, size_t src_cols, T* scale_vec, 
-        long int col_start, long int col_end ){
-     if(col_end < 0) col_end = src_cols + col_end ;  // change negative column index to positive
-     // scale each row (from offset index) in a matrix by a column vector
-     #ifdef _OPENMP
-     #pragma omp parallel for simd shared(src_mtx, src_rows, src_cols, scale_vec, col_start, col_end)
-     #endif
-     for(int i = 0; i< src_rows; i++){  
-          T scale = 1 / scale_vec[i];
-          for(int j=col_start; j<= col_end; j++){
-               src_mtx[src_cols*i + j] = src_mtx[src_cols*i + j] * scale ;
-          }
-     }
-}
-
-template<typename T>
-void norm_rows_by_maxabs_in_each_row(T* src_mtx, size_t src_rows, size_t src_cols, long int max_start_col, 
-        long int max_end_col, long int norm_start_col , long int norm_end_col){     
-     
-     T* norms = new T[src_rows]();     
-     get_max_each_row<T>(norms, src_mtx, src_rows, src_cols, max_start_col, max_end_col);   
-     norm_rows_in_mtx_by_col_vector<T>(src_mtx, src_rows, src_cols, norms, norm_start_col, norm_end_col);          
-     delete[] norms;
-}
-
-
-
-
-//Instantiate all non-template class/struct/method of specific type
-template bool init_mtx_in_mem<unsigned int>(unsigned int ** & data, size_t rows, size_t cols);
-
-template bool IsFloat<double>( std::string myString );
-template bool IsFloat<float>( std::string myString );
-
-template void clearMemo<double> (double ** & data);
-template void clearMemo<float>(float ** & data);
-template void clearMemo<unsigned int>(unsigned int ** & data);
-
-template void clearMemo<double>(std::vector<double**> & data);
-template void clearMemo<float>(std::vector<float**> & data);
-template void clearMemo<unsigned int>(std::vector<unsigned int**> & data);
-
-template void clearMemo<double>(std::map<std::string, double**> & data);
-template void clearMemo<float>(std::map<std::string, float**> & data);
-template void clearMemo<unsigned int>(std::map<std::string, unsigned int**> & data);
-
-template int read2DArrayfile<double>(double** & data, size_t& rows, size_t& cols, const char* file, int titleline=0);
-template int read2DArrayfile<float>(float** & data, size_t& rows, size_t& cols, const char* file, int titleline=0);
-
-template int read2DArray_with_max_thredhold<double>(double** & data, size_t& rows, size_t& cols, const char* file, int titleline, 
-                                            int thredhold_col, double thredhold_max);
-template int read2DArray_with_max_thredhold<float>(float ** & data, size_t& rows, size_t& cols, const char* file, int titleline, 
-                                            int thredhold_col, float thredhold_max);
-
-template void transpose_mtx<double>(double** & datdst,  double** datrsc,  size_t nrow_rsc, size_t ncol_rsc);
-template void transpose_mtx<float>(float** & datdst,  float** datrsc,  size_t nrow_rsc, size_t ncol_rsc);
-
-template void get_max_each_row<double>(double *& rst,  double* src,  size_t src_rows, size_t src_cols, 
-                        long int col_start, long int col_end); 
-template void get_max_each_row<float>(  float *& rst,   float* src,  size_t src_rows, size_t src_cols, 
-                        long int col_start, long int col_end); 
-
-template void norm_rows_in_mtx_by_col_vector<double>(double* src_mtx, size_t src_rows, size_t src_cols, double* scale_vec, 
-                        long int col_start=0, long int col_end=-1 );
-template void norm_rows_in_mtx_by_col_vector<float>(float* src_mtx, size_t src_rows, size_t src_cols, float* scale_vec, 
-                        long int col_start=0, long int col_end=-1 );
-
-template void norm_rows_by_maxabs_in_each_row<double>(double* src_mtx, size_t src_rows, size_t src_cols, long int max_start_col, 
-        long int max_end_col, long int norm_start_col , long int norm_end_col);
-template void norm_rows_by_maxabs_in_each_row<float>(float* src_mtx, size_t src_rows, size_t src_cols, long int max_start_col, 
-        long int max_end_col, long int norm_start_col , long int norm_end_col);
